add:画像の複数枚投稿を可能にした

+++ resources/views/musiris/index.blade.php
@@ -10,31 +10,39 @@
                 <h3 class='title'>
                     <a href="/post/{{ $post->id }}">{{ $post->title }}</a>
                 </h3>
-                <a href="/user/{{ $post->user->id }}">{{ $post->user->name }}</a>
-                <img src="{{ $post->image_url }}" alt="画像が読み込めません。"/>
-                <p>{{ $post->instruments }}</p>
+                <a class="userName" href="/user/{{ $post->user->id }}">{{ $post->user->name }}</a>
+                <div class="images">
+                    @foreach($post->images as $image)
+                     <img src="{{ $image->image_url }}" alt="画像が読み込めません。"/>
+                    @endforeach 
+                </div>
+                <p class="instruments">{{ $post->instruments }}</p>
                 <p>{{ $post->body }}</p>
             
-            <div class="button">
-                @if($post->islike($post))
-                    <div class="likes">
-                        <button class="like_Button liked" data-post-id="{{ $post->id }}">いいね解除</button>
-                    </div>
-                @else
-                    <div class="likes">
-                        <button class="like_Button" data-post-id="{{ $post->id }}">いいね</button>
-                    </div>
-                @endif
-                <div class="delete_Button">
-                <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}" method="post">
-                @csrf
-                @method('DELETE')
-                <button type="button" onclick="deletePost({{ $post->id }})">削除</button>
-                </form>
-                </div>
-            @endforeach
+                <div class="button">
+                    @if($post->islike($post))
+                        <div class="likes">
+                            <button class="like_Button liked" data-post-id="{{ $post->id }}">いいね解除</button>
+                        </div>
+                    @else
+                        <div class="likes">
+                            <button class="like_Button" data-post-id="{{ $post->id }}">いいね</button>
+                        </div>
+                    @endif
+                    @if($post->user->id == Auth::id())
+                        <div class="delete_Button">
+                            <form action="/posts/{{ $post->id }}" id="form_{{ $post->id }}" method="post">
+                            @csrf
+                            @method('DELETE')
+                            <button type="button" onclick="deletePost({{ $post->id }})">削除</button>
+                            </form>
+                        </div>
+                    @endif    
                 </div>
             </div>
+            @endforeach
+                
+            
     </div>
     <script>
     function deletePost(id){
